Fixed warnings and more
- MI::getInterrupt
- More SI::read regs
- Read VI_CONTROL

+++ src/ai.cpp
@@ -81,7 +81,7 @@ void AI::fillBuffer(uint32_t *out)
         {
             // Copy as many queued samples that can fit to the output buffer
             memcpy(&out[count / sizeof(uint32_t)], &buffer[offset], OUTPUT_SIZE - count);
-            offset += (OUTPUT_SIZE - count) / sizeof(uint32_t);
+            offset += (uint32_t)((OUTPUT_SIZE - count) / sizeof(uint32_t));
             return;
         }
     }
@@ -177,7 +177,7 @@ void AI::submitBuffer()
         // Copy samples to the buffer, scaled from their original frequency
         for (size_t i = 0; i < count; i++)
         {
-            uint32_t address = samples[0].address + (i * samples[0].count / count) * 4;
+            uint32_t address = samples[0].address + (uint32_t)(i * samples[0].count / count) * 4;
             uint32_t value = Memory::read<uint32_t>(0xA0000000 + address);
             buffer[i] = (value << 16) | (value >> 16);
         }
